Remove VOLUME directive in cvmfs-base
Exposing /cvmfs as a VOLUME in the base image seems to prevent
Dockerfiles of deriviative images from making persistent changes to this
directory. In particular, in cvmfs-atlas one gets an error at runtime
because the /cvmfs/atlas.cern.ch directory is missing, despite the fact
that it is created by the Dockerfile at build-time.